+++ examples/nlopt_policy/include/cartan_examples/nlopt/nlopt_bobyqa.h
@@ -204,7 +204,7 @@ public:
     void abort()
     {
         m_opt.force_stop();
-        m_status = ik_status::stalled;
+        m_status = ik_status::aborted;
     }
 
     /// Current solve policy status.