+++ lib/cartan-serial-chain/include/cartan/serial/ik/solver/filter_nw_sqp.h
@@ -248,7 +248,7 @@ public:
 
     void abort()
     {
-        m_status = ik_status::stalled;
+        m_status = ik_status::aborted;
         m_termination_reason = ik_termination_reason::solver_aborted;
     }